requirements.txt
africastalking==1.2.9
asgiref==3.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
dj-database-url==3.0.1
django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.3.0
psycopg==3.2.9
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.1.1
requests==2.32.4
rest-framework-simplejwt==0.0.2
schema==0.7.7
sqlparse==0.5.3
typing-extensions==4.14.1
urllib3==2.5.0
whitenoise==6.9.0
